requirements.txt
requests_oauthlib>=1.3.0
matplotlib>=3.5.0
scipy>=1.7.0
requests_cache>=1.0.0
numpy>=1.21.0
tqdm>=4.62.0
pandas>=1.3.0
attrs>=21.4.0
python-dotenv>=0.19.0
tenacity>=8.0.0
python-dotenv